abstract out tree policy
